requirements.txt
openai
youtube-transcript-api
python-dotenv
browser-use
playwright
python-dateutil
